ms-users/app/requirements.txt
Django==3.2.23
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
djangorestframework-simplejwt
